Resolvables refactoring
