+++ lib/hammer_cli_katello/activation_key.rb
@@ -6,7 +6,8 @@ module HammerCLIKatello
     resource :activation_keys
 
     class ListCommand < HammerCLIKatello::ListCommand
-      include KatelloEnvironmentNameResolvable
+      include LifecycleEnvironmentNameMapping
+      include LifecycleEnvironmentNameResolvable
       action :index
 
       output do
@@ -36,7 +37,8 @@ module HammerCLIKatello
     end
 
     class InfoCommand < HammerCLIKatello::InfoCommand
-      include KatelloEnvironmentNameResolvable
+      include LifecycleEnvironmentNameMapping
+      include LifecycleEnvironmentNameResolvable
       action :show
 
       def request_params
@@ -78,7 +80,8 @@ module HammerCLIKatello
     end
 
     class CreateCommand < HammerCLIKatello::CreateCommand
-      include KatelloEnvironmentNameResolvable
+      include LifecycleEnvironmentNameMapping
+      include LifecycleEnvironmentNameResolvable
       action :create
       success_message _("Activation key created")
       failure_message _("Could not create the activation key")
@@ -93,7 +96,8 @@ module HammerCLIKatello
     end
 
     class CopyCommand < HammerCLIKatello::CreateCommand
-      include KatelloEnvironmentNameResolvable
+      include LifecycleEnvironmentNameMapping
+      include LifecycleEnvironmentNameResolvable
       action :copy
 
       desc _("Copy an activation key")
@@ -106,7 +110,8 @@ module HammerCLIKatello
     end
 
     class UpdateCommand < HammerCLIKatello::UpdateCommand
-      include KatelloEnvironmentNameResolvable
+      include LifecycleEnvironmentNameMapping
+      include LifecycleEnvironmentNameResolvable
       action :update
       success_message _("Activation key updated")
       failure_message _("Could not update the activation key")
@@ -121,7 +126,8 @@ module HammerCLIKatello
     end
 
     class DeleteCommand < HammerCLIKatello::DeleteCommand
-      include KatelloEnvironmentNameResolvable
+      include LifecycleEnvironmentNameMapping
+      include LifecycleEnvironmentNameResolvable
       action :destroy
       success_message _("Activation key deleted")
       failure_message _("Could not delete the activation key")
