+++ lib/hammer_cli_katello/hostgroup_extensions.rb
@@ -13,7 +13,7 @@ module HammerCLIKatello
       def get_options(_defined_options, result)
         if result['option_environment_id'].nil? && result['option_environment_name']
           result['option_environment_id'] = @command.resolver.puppet_environment_id(
-            @command.resolver.scoped_options('environment', result))
+            @command.resolver.scoped_options('environment', result, :single))
         end
         result
       end
